wallet/signing: file renames
