+++ src/apps/wallet/sign_tx/segwit_bip143.py
